% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOTstyle.R
\name{scale_color_dot}
\alias{scale_color_dot}
\title{Function to access the custom color brewers of primary and secondary color palettes of DOT}
\usage{
scale_color_dot(palette = "main", discrete = TRUE, reverse = FALSE)
}
\arguments{
\item{palette}{Takes the values "primary" or "secondary" to access the combination of primary and secondary colors of DOT respectively}

\item{discrete}{Always set to the default value to be TRUE. Denotes that Discrete values are supplied to discrete scale. If changed to FALSE, errors out with the message "Discrete value supplied to continuous scale"}

\item{reverse}{Default value is False. If set to TRUE, reverses the color palette used}
}
\description{
This function takes two arguments namely "palette" and "reverse". Accesses the custom color palette with the combination of primary and secondary colors of DOT. Used in conjunction with ggplot. Used for plots to color "points" in a chart. Eg. scatterplots, residual plots etc
}
