% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOTtheme.R
\name{theme_dot}
\alias{theme_dot}
\title{Function to use DOT custom theme}
\usage{
theme_dot()
}
\arguments{
\item{No}{parameters are required}
}
\description{
Uses the theme_classic() as the default theme.
The Title font is Eurostar.
The Subtitle font is PT Sans.
The Caption (at the bottom) font is Arial.
}
